Add files via upload
3)PASSWORD BASED DOOR LOCK SYSTEM USING 8051 MICROCONTROLLER 
In our day to day lives, security is one of the prime concerns. We want to ensure safety to the fullest. Door control access is an important linkage in the security chain. A password-based door lock microcontroller is a system that permits access to only authorized users into secured/restricted areas.